test: combine tests into one spec for module
